:recycle: Refactor(admn/member): table 분리
Related to: #66
