+++ src/app/(main)/admin/_components/member/grade-table.tsx
@@ -1,29 +1,10 @@
 'use client'
 
-import { useState } from 'react'
-
-import {
-  ColumnDef,
-  ColumnFiltersState,
-  flexRender,
-  getCoreRowModel,
-  getFilteredRowModel,
-  getPaginationRowModel,
-  useReactTable,
-} from '@tanstack/react-table'
+import { ColumnDef } from '@tanstack/react-table'
 
 import { ManageUserGrageDB } from '@/lib/data'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/ui/table'
 
-import { TableFilter } from '../table-filter'
-import { TablePagination } from '../table-pagination'
+import { MemberTable } from '../member-table'
 import { GradeDialog } from './grade-dialog'
 
 const columns: ColumnDef<ManageUserGrade>[] = [
@@ -68,66 +49,5 @@ export const GradeTable = () => {
   // DB 연결 전 더미 데이터 사용
   const data: ManageUserGrade[] = ManageUserGrageDB
 
-  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([])
-  const [pagination, setPagination] = useState({
-    pageIndex: 0,
-    pageSize: 5,
-  })
-
-  const table = useReactTable({
-    data,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    onPaginationChange: setPagination,
-    onColumnFiltersChange: setColumnFilters,
-    state: {
-      pagination,
-      columnFilters,
-    },
-  })
-
-  const pageNumList = Array.from(
-    { length: table.getPageCount() },
-    (_, i) => i + 1,
-  )
-
-  return (
-    <div>
-      <TableFilter table={table} />
-      <Table>
-        <TableHeader>
-          {table.getHeaderGroups().map((headerGroup) => (
-            <TableRow key={headerGroup.id}>
-              {headerGroup.headers.map((header) => {
-                return (
-                  <TableHead key={header.id}>
-                    {header.isPlaceholder
-                      ? null
-                      : flexRender(
-                          header.column.columnDef.header,
-                          header.getContext(),
-                        )}
-                  </TableHead>
-                )
-              })}
-            </TableRow>
-          ))}
-        </TableHeader>
-        <TableBody>
-          {table.getRowModel().rows.map((row) => (
-            <TableRow key={row.id}>
-              {row.getVisibleCells().map((cell) => (
-                <TableCell key={cell.id}>
-                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                </TableCell>
-              ))}
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-      <TablePagination table={table} pageNumList={pageNumList} />
-    </div>
-  )
+  return <MemberTable data={data} columns={columns} />
 }
